+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Changelog
 
 0.2.6 (2017-01-?)
 ++++++++++++++++++
-* add delay param to bid method (#233)
+* add (minimum request) delay param (#233)
 * use Unauthorized expcetion (fix #232)
 
 0.2.5 (2016-12-28)